requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.1
django-cors-headers==3.13.0
django-markdownx==3.0.1
djangorestframework==3.13.1
filelock==3.7.1
Markdown==3.4.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.14.1
